work so far porting to windows

+++ external/tinycthread/tinycthread.c
@@ -151,8 +151,8 @@ int mtx_timedlock(mtx_t *mtx, const struct timespec *ts)
   }
   else
   {
-    timeoutMs  = (ts->tv_sec  - current_ts.tv_sec)  * 1000;
-    timeoutMs += (ts->tv_nsec - current_ts.tv_nsec) / 1000000;
+    timeoutMs  = (unsigned int)( (ts->tv_sec  - current_ts.tv_sec)  * 1000 );
+    timeoutMs += (unsigned int)( (ts->tv_nsec - current_ts.tv_nsec) / 1000000 );
     timeoutMs += 1;
   }
 
@@ -446,8 +446,9 @@ int cnd_timedwait(cnd_t *cond, mtx_t *mtx, const struct timespec *ts)
   struct timespec now;
   if (timespec_get(&now, TIME_UTC) == 0)
   {
-    DWORD delta = (ts->tv_sec - now.tv_sec) * 1000 +
-                  (ts->tv_nsec - now.tv_nsec + 500000) / 1000000;
+    DWORD delta = (unsigned int)
+				  ( (ts->tv_sec - now.tv_sec) * 1000 +
+                    (ts->tv_nsec - now.tv_nsec + 500000) / 1000000 );
     return _cnd_timedwait_win32(cond, mtx, delta);
   }
   else
@@ -701,9 +702,7 @@ int thrd_sleep(const struct timespec *duration, struct timespec *remaining)
 
   timespec_get(&start, TIME_UTC);
 
-  t = SleepEx(duration->tv_sec * 1000 +
-              duration->tv_nsec / 1000000 +
-              (((duration->tv_nsec % 1000000) == 0) ? 0 : 1),
+  t = SleepEx( (unsigned int) ( duration->tv_sec * 1000 + duration->tv_nsec / 1000000 + (((duration->tv_nsec % 1000000) == 0) ? 0 : 1) ),
               TRUE);
 
   if (t == 0) {
